More One-sided fixes:
  * Fix a counter roll-over issue that could result from a large (but
    not excessive) number of outstanding put/get/accumulate calls
    during a single synchronization issues (Refs trac:506)
  * Fix epoch issue with rdma component that would effect PWSC
    synchronization (Refs trac:507)

This commit was SVN r12673.

The following Trac tickets were found above:
  Ticket 506 --> https://svn.open-mpi.org/trac/ompi/ticket/506
  Ticket 507 --> https://svn.open-mpi.org/trac/ompi/ticket/507

+++ ompi/mca/osc/pt2pt/osc_pt2pt_component.c
@@ -257,7 +257,7 @@ ompi_osc_pt2pt_component_select(ompi_win_t *win,
     OBJ_CONSTRUCT(&module->p2p_pending_control_sends, opal_list_t);
 
     OBJ_CONSTRUCT(&module->p2p_pending_sendreqs, opal_list_t);
-    module->p2p_num_pending_sendreqs = (short*)malloc(sizeof(short) * 
+    module->p2p_num_pending_sendreqs = (unsigned int*)malloc(sizeof(unsigned int) * 
                                                       ompi_comm_size(module->p2p_comm));
     if (NULL == module->p2p_num_pending_sendreqs) {
         OBJ_DESTRUCT(&module->p2p_pending_sendreqs);
@@ -268,7 +268,7 @@ ompi_osc_pt2pt_component_select(ompi_win_t *win,
         return ret;
     }
     memset(module->p2p_num_pending_sendreqs, 0, 
-           sizeof(short) * ompi_comm_size(module->p2p_comm));
+           sizeof(unsigned int) * ompi_comm_size(module->p2p_comm));
 
     module->p2p_num_pending_out = 0;
     module->p2p_num_pending_in = 0;
@@ -279,7 +279,7 @@ ompi_osc_pt2pt_component_select(ompi_win_t *win,
     OBJ_CONSTRUCT(&(module->p2p_long_msgs), opal_list_t);
 
     OBJ_CONSTRUCT(&(module->p2p_copy_pending_sendreqs), opal_list_t);
-    module->p2p_copy_num_pending_sendreqs = (short*)malloc(sizeof(short) * 
+    module->p2p_copy_num_pending_sendreqs = (unsigned int*)malloc(sizeof(unsigned int) * 
                                                            ompi_comm_size(module->p2p_comm));
     if (NULL == module->p2p_copy_num_pending_sendreqs) {
         OBJ_DESTRUCT(&module->p2p_copy_pending_sendreqs);
@@ -293,7 +293,7 @@ ompi_osc_pt2pt_component_select(ompi_win_t *win,
         return ret;
     }
     memset(module->p2p_num_pending_sendreqs, 0, 
-           sizeof(short) * ompi_comm_size(module->p2p_comm));
+           sizeof(unsigned int) * ompi_comm_size(module->p2p_comm));
 
     /* fence data */
     module->p2p_fence_coll_counts = (int*)malloc(sizeof(int) * 
@@ -314,7 +314,7 @@ ompi_osc_pt2pt_component_select(ompi_win_t *win,
         module->p2p_fence_coll_counts[i] = 1;
     }
 
-    module->p2p_fence_coll_results = (short*)malloc(sizeof(short) * 
+    module->p2p_fence_coll_results = (unsigned int*)malloc(sizeof(unsigned short) * 
                                                     ompi_comm_size(module->p2p_comm));
     if (NULL == module->p2p_fence_coll_results) {
         free(module->p2p_fence_coll_counts);
